requirements.txt
fastapi~=0.115.5
numpy~=2.1.3
pillow~=11.0.0
uvicorn~=0.32.1
python-multipart
opencv-python~=4.10.0.84
scikit-learn~=1.5.2
plotly~=5.24.1
matplotlib~=3.9.2
scikit-spatial
sympy
scipy~=1.14.1
scikit-image~=0.24.0
torch~=2.5.1
python-dotenv~=1.0.1
